Captain Range bubble
Added the script for "2nd captain" which allows for captain cards with a Range bubble. Based on the script for "2nd weapon" which adds the Range bubble for weapon upgrades that use them.

Created two new ranged faction templates. These essentially are weapons templates with the Attack value box removed, but the Range bubble left in place. (Overlayed the top of a non-weapon card over the top of the weapon template.)